Shell specific path command when installing `shm`
Previously `shm` was using a naive solution of `.profile` to use a
standardized way to add .shm directory to user path. The `.profile` file
unfortunately does not run, when opening new shells.

To overcome this the user's current shell is evaluated, matched and
approriate set path command is applied to the shell's configuration
file. Specific support for fish, zsh and bash. All the other shells will
add the set path to `.profile` file as before.

If shm directory is found under user's shell configuration, it will not
be appended again. Without this check multiple appends could happen in a
scenario where user has removed .shm directory, but not the pathing
configuration. Multiple path commands for the same variable might result
in a panic (like in case with fish shell).