Use $HOME variable instead of literal path
Previously a new installation of `.shm` would write a literal path to
user's home directory. This made the installation unusable across
different machines due to possibility of different path for user home
directory.

In this implementation the problem is addressed by using the $HOME
variable to decide where the shm installed packages are located.