Add `add` command to add local scripts to `shm`
Purpose of this command is to provide quick and easy way to test local
scripts with the ease of shm already being available in the $PATH.

With the add command `shm` can now make simple local scripts globally
available for the user.

Add command provides a flag `-f` or `--force` to overwrite any existing
scripts.