Add remapping for x86_64 to support amd64 term
This should fix the exit issue when using intel based macs. The grep
didn't find a match and silently failed since x86_64 might not be found
in the release asset.

Also support x86_64 when the mac has arm64 architecture as some release
assets only provide the x86_64 version for arm macs took. I'm not aware
of any issues using binaries made for that architecture as of now.